refactor: 컴포넌트 이름 명확하게 변경
