+++ src/components/ClassGenerateButton/index.tsx
@@ -1,8 +1,8 @@
 import Class from "@/assets/class_natural-500.svg"
 
-import { Base } from "@/components/ClassGenerate/style";
+import { Base } from "@/components/ClassGenerateButton/style";
 
-export default function ClassGenerate() {
+export default function ClassGenerateButton() {
   return (
     <Base>
       <img src={Class} alt="Docs"/>클래스 개설하기</Base>
