+++ src/components/ClassGenerateButton/style.ts
