+++ src/components/RecruitmentDetailButton/index.tsx
@@ -1,9 +1,9 @@
 /** @jsxImportSource @emotion/react */
 import Rarrow from "@/assets/Rarrow_natural-500.svg";
 
-import { Base } from "@/components/RecruitmentDetail/styles";
+import { Base } from "@/components/RecruitmentDetailButton/styles";
 
-export default function Detail() {
+export default function RecruitmentDetailButton() {
   return (
     <Base>
       모집 글 자세히 보기
